feat(AuxTimeline): Add Auxiliary Timeline core
The purpose of aux timelines is for second storylines. Aux Timelines are created using the same process as mission timelines, but are designated as aux timelines in the Mission Information section. You can run as many aux timelines as you want and switch between them using the aux timeline core. Other than that, they work exactly the same as regular timelines.